pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dash>=3.3.0",
    "pyinstaller>=6.17.0",
    "pyyaml>=6.0.3",
]
